Multiple Parameter Lists (#227)
* Generalize `Apps`

Previsouly `Apps` unapplies the tree of form `App(App(App(id: Ident, p1), p2), p3)` to `Apps(id: Ident, p1 :: p2 :: p3 :: Nil)`. Now it unapplies this as well but also `App(App(App(tree: Tree, p1), p2), p3)` to `Apps(tree: Tree, p1 :: p2 :: p3 :: Nil)`. This is especially useful for type parameters as we want to unwarp as many of layers of `App` as we can, then unwrap the `TyApp`.

* Generalize `TermDef` for multiple parameter lists

The member `params` is renamed to `paramLists` to reflect its new behavior. The type of `paramLists` is changed from `Opt[Ls[Tree]]` to `Ls[Tup]` to reflect its new behavior. It now returns a list of parameter tuples instead of a list of parameters.

* Adapt `Elaborator` and `JSBuilder` for multiple parameter lists

* Refactor `GlobalThis.mls` to remove machine-dependent contents from diff-test output

---------

Co-authored-by: Lionel Parreaux <lionel.parreaux@gmail.com>